index 수정 0213 오전본

+++ food_hub_project/src/main/java/com/application/foodhub/post/PostDAO.java
@@ -42,5 +42,5 @@ public interface PostDAO {
 	public String getCategoryName(@Param("categoryId") Long categoryId);
 
 	// 카테고리별 최신글 2개 가져오기
-	public List<Map<String, Object>> getLatestPostsByCategoryId(@Param("categoryId") int categoryId);
+	public List<Map<String, Object>> getLatestPostsByCategoryId(@Param("categoryId") long categoryId);
 }
