+++ food_hub_project/src/main/java/com/application/foodhub/post/PostServiceImpl.java
@@ -109,7 +109,7 @@ public class PostServiceImpl implements PostService {
 	}
 
 	@Override
-	public List<Map<String, Object>> getLatestPostsByCategoryId(int categoryId) {
+	public List<Map<String, Object>> getLatestPostsByCategoryId(long categoryId) {
 		return postDAO.getLatestPostsByCategoryId(categoryId);
 	}
